[IMP] web: allow executing an action in a new window
This commit adds a new mechanism to be able to execute an action in a
new window. This is useful, for instance, to open records in a new
window. Note that, the action is stored in the session storage, before
opening the new window. The new window will have a copy of the "parent"
window (see : [1] "After creating a new auxiliary browsing context and
document, the session storage is copied over.".
After opening the new window, the action is immediately removed from the
session storage.

task-id 4031939

[1] : https://html.spec.whatwg.org/multipage/webstorage.html#webstorage)

Co-authored-by: Luca Vitali <luvi@odoo.com>
Co-authored-by: Jorge Pinna Puissant <jpp@odoo.com>